Move StatsPage component into self-contained folder
