+++ src/app.tsx
@@ -7,7 +7,7 @@ import { TaskInput } from './components/task-input/task-input'
 import { Filters } from './components/filters/filters'
 import { TaskList } from './components/task-list/task-list'
 import { Footer } from './components/footer/footer'
-import { StatsPage } from './components/stats-page'
+import { StatsPage } from './components/stats-page/stats-page'
 import { NavLink } from './components/nav-link/nav-link'
 
 export function App() {
